Refactor namespaces in ColorMapEditor project
The namespaces in the files of the ColorMapEditor project have been updated for better organization and readability. The specific changes include moving FrmMain and ColorHeatMap to the Forms and Classes namespaces, respectively, while also introducing necessary 'using' directives.
