+++ ColorMapEditor/Classes/ColorHeatMap.cs
@@ -4,7 +4,7 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 
-namespace ColorMapEditor
+namespace ColorMapEditor.Classes
 {
     public class ColorHeatMap
     {
