+++ ColorMapEditor/Controls/Table3DView.cs
@@ -4,6 +4,7 @@ using System.Drawing;
 using System.Globalization;
 using System.Linq;
 using System.Windows.Forms;
+using ColorMapEditor.Classes;
 
 namespace ColorMapEditor.Controls
 {
