+++ ColorMapEditor/Forms/FrmMain.cs
@@ -1,15 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
+using ColorMapEditor.Classes;
 
-namespace ColorMapEditor
+namespace ColorMapEditor.Forms
 {
     public partial class FrmMain : Form
     {
